requirements.txt
asgiref==3.8.1
Django==5.2
django-jinja==2.11.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
sqlparse==0.5.3
django-debug-toolbar==4.4.6
yandex-cloud-ml-sdk==0.9.1
yandexcloud==0.341.0
yarl==1.20.0
python-dotenv==1.0.0
djangorestframework>=3.14
drf-spectacular>=0.27
PyYAML>=6.0
uritemplate>=4.1
drf-spectacular>=0.27,<1.0
djangorestframework>=3.14,<4.0
drf-yasg>=1.21,<2.0